Move tests around and clarify naming
